EC2-Crawler/requirements.txt
boto3
PyYAML
requests
beautifulsoup4